Code refactoring
- create folder for internal loaders and plugins
- fix the length of the lines
- update readme files
